+++ packages/terra-framework-docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Changed
   * Updated `flowsheet-data-grid` tests and examples to support non contiguous multi select.
+  * Updated `terra-date-input-field` example to have visual label and contextual error messages.
 
 ## 1.54.0 - (January 2, 2024)